refactor(terminals): use single files for single-file adapters
Revert directory structure for adapters that only have one file.
Only wezterm keeps a subdirectory since it has multiple files (lua.py).

This simplifies the codebase while maintaining consistent import paths:
- from .alacritty import AlacrittyAdapter  (single file)
- from .wezterm import WeztermAdapter      (package with __init__.py)

Both patterns work identically from the consumer's perspective.
