+++ console_cowboy/terminals/alacritty.py
@@ -34,7 +34,7 @@ from console_cowboy.ctec.schema import (
 )
 from console_cowboy.utils.colors import normalize_color
 
-from ..base import TerminalAdapter
+from .base import TerminalAdapter
 
 
 class AlacrittyAdapter(TerminalAdapter):
