--- console_cowboy/terminals/alacritty/__init__.py
@@ -1,5 +1,0 @@
-"""Alacritty terminal adapter."""
-
-from .adapter import AlacrittyAdapter
-
-__all__ = ["AlacrittyAdapter"]
