--- console_cowboy/terminals/ghostty/__init__.py
@@ -1,5 +1,0 @@
-"""Ghostty terminal adapter."""
-
-from .adapter import GhosttyAdapter
-
-__all__ = ["GhosttyAdapter"]
