--- console_cowboy/terminals/iterm2/__init__.py
@@ -1,5 +1,0 @@
-"""iTerm2 terminal adapter."""
-
-from .adapter import ITerm2Adapter
-
-__all__ = ["ITerm2Adapter"]
