--- console_cowboy/terminals/kitty/__init__.py
@@ -1,5 +1,0 @@
-"""Kitty terminal adapter."""
-
-from .adapter import KittyAdapter
-
-__all__ = ["KittyAdapter"]
