--- console_cowboy/terminals/terminal_app/__init__.py
@@ -1,5 +1,0 @@
-"""macOS Terminal.app adapter."""
-
-from .adapter import TerminalAppAdapter
-
-__all__ = ["TerminalAppAdapter"]
