+++ console_cowboy/terminals/vscode.py
@@ -26,7 +26,7 @@ from console_cowboy.ctec.schema import (
 )
 from console_cowboy.utils.colors import normalize_color
 
-from ..base import TerminalAdapter
+from .base import TerminalAdapter
 
 
 class VSCodeAdapter(TerminalAdapter):
