--- console_cowboy/terminals/vscode/__init__.py
@@ -1,5 +1,0 @@
-"""VS Code terminal adapter."""
-
-from .adapter import VSCodeAdapter
-
-__all__ = ["VSCodeAdapter"]
